backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.20
langgraph==0.0.40
langchain-openai==0.0.8
langchain-community==0.0.20
tavily-python==0.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

